+++ tests/util.py
@@ -10,6 +10,7 @@ class TestBase(unittest.TestCase):
     def setUp(self):
         import warnings
         warnings.simplefilter("ignore", ResourceWarning)
+        warnings.simplefilter("ignore", SyntaxWarning)
         surpress_warnings()
         self.text_1 = 'He was diagnosed with kidney failure and heart disease.'
         self.text_2 = 'He loved to smoke but Marlboro cigarettes gave John Smith lung cancer while he was in Chicago.'